Add a simple operator to apply a "counter 1/f" filter. (#863)
* Add a simple operator to apply a "counter 1/f" filter.

For (at least) 25 years, the inverse time domain noise covariance
has been used as a "change of basis" when solving a generalized
least squares problem for the maximum likelihood map in pixel space.

For simpler "filter and bin" mapmaking techniques, application of
this N_tt'^-1 filter may be useful.  This work adds an operator
which uses a given noise model for detector PSDs to compute the
inverse noise covariance fourier domain kernels for each detector
and convolve the timestreams by those.

* Add initial unit test.  Fix normalization.

* More unit test improvements and cleanup